feat(price): 해외 실시간 시세·거래대금 랭킹 - Toss 신규 API 기반
ROADMAP.md §10(해외 실시간 시세)·§2c(거래대금 실시간 랭킹)를 원래
KIS 웹소켓으로 설계했으나, toss-openapi.json 최신 버전(1.2.5)에
prices/candles의 해외 티커 지원과 신규 rankings 엔드포인트가 이미
있는 걸 라이브 호출로 확인해 계획을 전면 재설계했다. KIS 웹소켓
(승인키·AES·PINGPONG·재연결 상태기계) 없이 기존 Toss 폴링 인프라를
그대로 확장한다.

- infra/toss: TossApiClient.getRankings()/getUsMarketCalendar() 추가
  (기존 withTokenRetry+ExternalApiInvoker 패턴 재사용)
- market: TossMarketRankingCache 신규 - Toss /api/v1/rankings 기반
  국내/해외 공용 랭킹 캐시(10초 TTL). MarketRankingService가 scope
  (domestic|overseas)/sort(gainers|losers|amount|volume)를 이 캐시로
  라우팅하되, "관심종목만 보기"+gainers/losers는 top100 제약을 피하기
  위해 기존 자체 계산 경로(국내 MarketRankingCache, 해외 신규
  OverseasPreviousCloseCache)를 그대로 유지
- price: OverseasWatchlistPriceScheduler 신규 - 관심종목의 해외 코드만
  폴링해 Redis 적재+STOMP 즉시 브로드캐스트. UsMarketCalendarCache로
  게이팅(미국 정규장이 KST 자정을 넘기는 문제를 3영업일 세션 전체
  캐싱으로 처리)
- price: WatchlistedStockCodeCache를 국내 전용으로 좁힘(신규 해외
  스케줄러와의 중복 브로드캐스트 방지), ChangeRateCalculator 공용
  유틸 추출(3곳에 중복돼 있던 등락률 계산 통합)
- price: StockPriceService.getCurrentPrice 캐시 미스 시 해외 종목은
  OverseasDailyPriceRepository로 폴백하도록 분기. 캐시 히트 경로가
  시장 구분과 무관하게 항상 "KRW"를 반환하던 기존 버그도 함께 수정
- PriceSnapshot/CurrentPriceResponse/MarketRankingResponse의 현재가
  필드를 Long→Double로 확대(해외 USD 소수점 지원)
- 프론트: RankingTable의 해외 플레이스홀더 제거하고 scope 실제 전달,
  통화 인지 formatPrice로 관심종목/상세 페이지 USD 표시

검증: 실제 크레덴셜로 라이브 API 5건(현재가 해외/혼합, 캔들 해외,
랭킹 KR/US) 확인 후 착수. 백엔드 386개 테스트(신규 38개 포함) 전부
통과, 프론트 tsc -b + oxlint 통과.

+++ backend/core/src/main/java/com/quantlime/price/dto/response/CurrentPriceResponse.java
@@ -1,8 +1,12 @@
 package com.quantlime.price.dto.response;
 
+/**
+ * {@code price}는 2026-07-29 해외(USD) 실시간가 지원을 위해 Long에서
+ * Double로 확대됐다.
+ */
 public record CurrentPriceResponse(
     String stockCode,
-    Long price,
+    Double price,
     Double changeRate,
     String currency,
     String timestamp
